Add files via upload
Version of race.asm that can run on FliP's Multicart. The multicart does not map $a00-$bff to the cartridge, so the game was remapped to use $100-$7ff and $c00-$fff. Pointers and interrupt is initialised from $000 onwards - the rest of the kernel is left out. Interestingly, this leaves more space for game expansion (though this may need additional shuffling $400-$4ff is currntly not used).